+++ project/Dependencies.scala
@@ -68,9 +68,11 @@ object Dependencies {
       "ch.qos.logback" % "logback-classic" % logbackVersion,
       "com.typesafe.scala-logging" %% "scala-logging" % loggingVersion
     )
-    val kryo = "com.esotericsoftware" % "kryo" % kryoVersion
-    val xgboostDep = Seq("ml.dmlc" %% "xgboost4j" % xgboostVersion exclude("com.esotericsoftware.kryo", "kryo"), kryo)
-    val xgboostSparkDep = Seq("ml.dmlc" %% "xgboost4j-spark" % xgboostVersion exclude("com.esotericsoftware.kryo", "kryo"), kryo)
+    val kryo = Seq(ExclusionRule("com.esotericsoftware.kryo", "kryo"),  "com.esotericsoftware" % "kryo" % kryoVersion)
+    val xgboostDep = "ml.dmlc" %% "xgboost4j" % xgboostVersion
+    val xgboostPredictorDep = "biz.k11i" % "xgboost-predictor" % "0.3.1"
+    val xgboostSparkDep = "ml.dmlc" %% "xgboost4j-spark" % xgboostVersion
+
     val hadoop = "org.apache.hadoop" % "hadoop-client" % hadoopVersion
   }
 
@@ -118,9 +120,9 @@ object Dependencies {
 
   val tensorflow = l ++= tensorflowDeps ++ Seq(Test.scalaTest)
 
-  val xgboostRuntime = l ++= xgboostDep ++ Test.spark ++ Seq(Test.scalaTest)
+  val xgboostRuntime = l ++= Seq(xgboostDep) ++ Seq(xgboostPredictorDep) ++ Seq(kryo) ++ Test.spark ++ Seq(Test.scalaTest)
 
-  val xgboostSpark = l ++= xgboostSparkDep ++ Provided.spark
+  val xgboostSpark = l ++= Seq(xgboostSparkDep) ++ Seq(kryo) ++ Provided.spark
 
   val serving = l ++= Seq(akkaHttp, akkaHttpSprayJson, config, Test.scalaTest, Test.akkaHttpTestkit)